regrename: Don't rename the dest of a REG_CFA_REGISTER (PR85645)
We should never change the destination of a REG_CFA_REGISTER, just
like for insns with a REG_CFA_RESTORE, because we need to have the
same control flow information on all branches that join.  It is very
doubtful that renaming the scratch registers used for prologue/epilogue
will help anything either.


	PR rtl-optimization/85645
	* regrename.c (build_def_use): Also kill the chains that include the
	destination of a REG_CFA_REGISTER note.


git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@260075 138bc75d-0d04-0410-961f-82ee72b054a4