[arm] PR target/85733 Restore be8 linking behaviour for ARMv6-M and products deriving from its capabilities
My patch last year to automate passing the be8 flag to the linker had
a nasty flaw in that I forgot entirely that the ARMv6-M architecture
did not derive its capabilities directly from the ARMv6 capability
list, but was a new group of capabilities (since it needs to leave out
the ARM -- notm -- feature bit).  The feature list defined was thus
missing the be8 bit.  Furthermore, any product derived from that
feature group consequently lacked the be8 feature as well and this
included all ARMv7 and ARMv8 parts.

The fix is embarrassingly simple...

	PR target/85733
	* config/arm/arm-cpus.in (fgroup ARMv6m): Add be8 feature.


git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@260162 138bc75d-0d04-0410-961f-82ee72b054a4